eye_exam/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
collection==0.1.6
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
Django==4.0.1
django-admin-interface==0.18.4
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-nested-admin==3.4.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.30.0
idna==3.3
jsonpickle==2.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.0
pycparser==2.21
pyparsing==3.0.7
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.3
pyzbar==0.1.8
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
transliterate==1.10.2
tzdata==2021.5
urllib3==1.26.8
